requirements.txt
fpdf==1.7.2
Pillow==8.1.0
PyQRCode==1.2.1
python-barcode==0.13.1
